requirements.txt
amqp==1.4.6
anyjson==0.3.3
asyncio==3.4.3
billiard==3.3.0.20
celery==3.1.18
Django==1.8.1
django-celery==3.1.16
django-countries==3.3
foursquare===1-2015.04.07
httplib2==0.9.1
kombu==3.0.26
numpy==1.8.2
pygobject==3.10.2
python-instagram==1.3.1
python3-instagram==0.9.0
pytz==2015.2
requests==2.7.0
simplejson==3.6.5
six==1.9.0
virtualenv==12.1.1
